requirements.txt
streamlit>=1.28.0
openai>=1.0.0
scikit-learn>=1.3.0,<1.4.0
numpy>=1.21.0,<1.25.0
plotly>=5.15.0
python-dotenv>=1.0.0
pandas>=1.5.0,<2.1.0
matplotlib>=3.6.0,<3.8.0
scipy>=1.9.0,<1.11.0
